test: move updater script fixtures into fixtures module
